+++ index_sse2.h
@@ -1,6 +1,8 @@
 #ifndef CSIMDBYTE_INDEX_SSE2_H
 #define CSIMDBYTE_INDEX_SSE2_H
 
-size_t index_sse2(const char *s, size_t sz, const char *ss, size_t ssz);
+#include <stdint.h>
+
+int64_t index_sse2(const char *s, int64_t sz, const char *ss, int64_t ssz);
 
 #endif //CSIMDBYTE_INDEX_SSE2_H
